Optimized levelup max algorithm
Can compute max reachable level with current balance using the convergence of the function f(x) = ( sum of (n^2) from n=1 to n=x ) where f(b*x)/f(x) ≈ b^3 for large x.
In simple terms, if you have spent a total of x coins to reach a level y, then it takes a total of 8x coins to reach a level 2y. This means that you need a balance of 7x coins (you already spent x many coins, 8x is the total) to level up y times (from y to 2y). (granted that x and y are sufficiently large.)
However, a perfect convergence takes an obscenely high level, so this algorithm might give a few too many / a few too little levels, which might set the balance negative, so I made it a separate command so players know what they're doing. If it works as intended then it can be rolled into the main command.
Players will always be charged the appropriate amount for their levelups (see the formula in Calc.js, plug in any values you want) so it is not exploitable.
There are also NaN checks to prevent data corruption.